+++ CHANGELOG.md
@@ -21,3 +21,4 @@
 
 ### Back-end
 - Reworked entity system to accept more model entities
+- Automatically disconnect from the RFU Back-end when AFK